# webui/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CodeExploder.Gateway | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/sessions": {
      "get": {
        "tags": [
          "CodeExploder.Gateway"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CodeExploder.Gateway"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionSummary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}": {
      "get": {
        "tags": [
          "CodeExploder.Gateway"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionSummary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "CodeExploder.Gateway"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/sessions/{id}/analysis": {
      "get": {
        "tags": [
          "CodeExploder.Gateway"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisSnapshot"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisSnapshot": {
        "required": [
          "status",
          "stages",
          "narration",
          "lastEventId",
          "summary"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "stages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StageInfo"
            }
          },
          "narration": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NarrationLine"
            }
          },
          "lastEventId": {
            "type": "integer",
            "format": "int64"
          },
          "summary": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RepoSummary"
              }
            ]
          }
        }
      },
...
      "ChurnStat": {
        "required": [
          "path",
          "commits"
        ],
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "commits": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ComponentSummary": {
        "required": [
          "name",
          "fileCount",
          "topFiles"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "fileCount": {
            "type": "integer",
            "format": "int32"
          },
          "topFiles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "LanguageStat": {
        "required": [
          "name",
          "files",
          "bytes",
          "percent"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "files": {
            "type": "integer",
            "format": "int32"
          },
          "bytes": {
            "type": "integer",
            "format": "int64"
          },
          "percent": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "NarrationLine": {
        "required": [
          "at",
          "text"
        ],
        "type": "object",
        "properties": {
          "at": {
            "type": "string",
            "format": "date-time"
          },
          "text": {
            "type": "string"
          }
        }
      },
...
      "NewSessionRequest": {
        "required": [
          "url",
          "gitRef"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "gitRef": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RepoSummary": {
        "required": [
          "commitSha",
          "description",
          "fileCount",
          "analyzedFileCount",
          "excludedFileCount",
          "chunkCount",
          "totalBytes",
          "languages",
          "buildSystems",
          "ciConfigs",
          "entryPoints",
          "components",
          "topChurnFiles",
          "commitCount",
          "contributorCount"
        ],
        "type": "object",
        "properties": {
          "commitSha": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "fileCount": {
            "type": "integer",
            "format": "int32"
          },
          "analyzedFileCount": {
            "type": "integer",
            "format": "int32"
          },
          "excludedFileCount": {
            "type": "integer",
            "format": "int32"
          },
          "chunkCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalBytes": {
            "type": "integer",
            "format": "int64"
          },
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LanguageStat"
            }
          },
          "buildSystems": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ciConfigs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "entryPoints": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "components": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComponentSummary"
            }
          },
          "topChurnFiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChurnStat"
            }
          },
          "commitCount": {
            "type": "integer",
            "format": "int32"
          },
          "contributorCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "SessionProgress": {
        "required": [
          "completedSections",
          "totalSections"
        ],
        "type": "object",
        "properties": {
          "completedSections": {
            "type": "integer",
            "format": "int32"
          },
          "totalSections": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SessionSummary": {
        "required": [
          "id",
          "kind",
          "title",
          "repoOwner",
          "repoName",
          "prNumber",
          "status",
          "failureReason",
          "createdAt",
          "progress",
          "analysisId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "repoOwner": {
            "type": "string"
          },
          "repoName": {
            "type": "string"
          },
          "prNumber": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "failureReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "progress": {
            "$ref": "#/components/schemas/SessionProgress"
          },
          "analysisId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "StageInfo": {
        "required": [
          "key",
          "label",
          "state",
          "percent",
          "detail"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "percent": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}